requirements.txt
asgiref==3.4.1
Django~=3.2.7
django-neomodel==0.0.7
djangorestframework==3.12.4
wheel~=0.36.2
neo4j-driver==4.3.1
neobolt==1.7.17
neomodel==4.0.6
psycopg2==2.9.1
pytz==2021.1
Shapely==1.7.1
six==1.16.0
sqlparse~=0.4.2
connection~=2021.7.20
cryptography~=3.4.7
tornado~=6.1
colorama~=0.4.4
Jinja2~=2.11.3
numpy~=1.20.3
Cython~=0.29.22
requests~=2.25.1
Pygments~=2.8.1
django-cors-headers~=3.11.0
selenium~=4.1.0
torch==1.8.2+cpu 
torchvision==0.9.2+cpu 
torchaudio===0.8.2
scikit-learn==1.0.2
